requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.2

# Database & Storage
appwrite==4.1.0

# Network utilities
wakeonlan==3.1.0

# Performance and monitoring
psutil==5.9.6
gunicorn==21.2.0

# HTTP client improvements (optional but recommended)
httpx==0.25.2
aiofiles==23.2.1

# Production dependencies
prometheus-fastapi-instrumentator==6.1.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Base dependencies from original requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
idna==3.6
ipykernel==6.26.0
ipython==8.18.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12